Refactor SpringReference (#4805)
* Separating and clearer naming

* Add license header

* Review feedback
