+++ rewrite-xml/src/main/resources/META-INF/services/org.openrewrite.trait.Reference$Provider
@@ -1,1 +1,1 @@
-org.openrewrite.xml.trait.SpringReference$Provider
+org.openrewrite.xml.trait.SpringXmlReference$Provider
