+++ rewrite-xml/src/test/java/org/openrewrite/xml/trait/SpringXmlReferenceTest.java
@@ -23,11 +23,11 @@ import org.openrewrite.test.RewriteTest;
 
 import static org.openrewrite.xml.Assertions.xml;
 
-class SpringReferenceTest implements RewriteTest {
+class SpringXmlReferenceTest implements RewriteTest {
 
     @Override
     public void defaults(RecipeSpec spec) {
-        spec.recipe(RewriteTest.toRecipe(() -> new SpringReference.Matcher()
+        spec.recipe(RewriteTest.toRecipe(() -> new SpringXmlReference.Matcher()
           .asVisitor(springJavaTypeReference -> SearchResult.found(springJavaTypeReference.getTree(), springJavaTypeReference.getValue()))));
     }
